pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "relate-courseware"
version = "2024.1"
description = "RELATE courseware"
authors = [{ name = "Andreas Kloeckner", email = "inform@tiker.net" }]
requires-python = ">=3.11,<4"
readme = "README.rst"
license = "MIT"
dependencies = [
    # upgrade django-stubs below when bumping this
    "django>=5.2,<6",

    "django-crispy-forms>=1.13.0",
    "crispy-bootstrap5~=2025.4",
    "jsonfield>=1.4.0",
    "django-yamlfield",
    "django-stubs-ext",
    "markdown~=3.7",
    "minijinja>=2.7.0,<3",
    "pymbolic>=2024.2.2",
    "sympy>=1.13.3",
    "pyyaml>=6.0.2",
    "dulwich>=0.22.6,<0.24",
    "paramiko>=3.5.1",
    "slixmpp>=1.8.3,<2",
    "docker>=7.1.0,<8",
    "pygments>=2.6.1,<3",
    "unicodecsv>=0.14.1,<0.15",
    "pysaml2>=7.5.2,<8",
    "djangosaml2>=1.7.0,<2",
    "attrs>=19",
    "python-memcached~=1.59",
    "celery>=5.2.2,<6",
    "kombu>=5.4.2,<6",
    "django-celery-results>=2.4.0,<3",
    "django_select2>=8.2.1,<9",
    "bleach~=6.2",
    "html5lib~=1.1",
    "pytools>=2024.1.8",
    "colorama>=0.4.6",
    "cryptography>=3.2.1",
    "social-auth-app-django>=5.4.1,<6",
    "urllib3>=2.3.0,<3",
    "typing_extensions>=4.14.1",

    # "pydantic~=2.11.7",
    # https://github.com/pydantic/pydantic/pull/12279
    # also diallow direct references below when reverting
    "pydantic @ git+https://github.com/inducer/pydantic.git@patch-1",

    # needed by pydantic EmailStr
    "email-validator~=2.3.0",
    "annotated-types~=0.7.0",

    "packaging>=25",

    # prevent spurious downgrade by uv? Counters upper bound by pysaml2:
    # https://github.com/IdentityPython/pysaml2/blob/0252ec96058c87c43f89e05d978e72b6ba2e6978/CHANGELOG.md?plain=1#L10
    "pyopenssl>=24.2.1"
]

[project.optional-dependencies]
postgres = ["psycopg2>=2.9.10,<3"]
memcache = ["pylibmc>=1.6.0,<2"]

[project.urls]
Repository = "https://github.com/inducer/relate"
Documentation = "https://documen.tician.de/relate/"

[project.scripts]
relate = "relate.bin.relate:main"

[dependency-groups]
dev = [
    "factory_boy>=3.3.1,<4",
    "ruff>=0.12.3,<1",
    "django-types",
    "basedpyright>=1.30.1",
    "pytest>=8,<9",
    "pytest-django>=4.5.2,<5",
    "pytest-factoryboy>=2.6.0,<3",
    "jsonpatch>=1.33",

    # https://github.com/wronglink/pytest-pudb/pull/29
    "pytest_pudb @ git+https://github.com/ndhansen/pytest-pudb.git@c29b32a7c93e9a2d40ae1d8594c1fcb3dfd26e06",

    "safety>=3.2.11,<4",
    # workaround for:
    # https://github.com/pyupio/safety/issues/784
    "typer<0.20",

    "sphinx>=8.1.3,<9",
    "furo>=2024.8.6,<2026",
    "sphinx-copybutton>=0.5.2,<0.6",
    "django-upgrade>=1.22.1,<2",
    "numpy~=2.1",
    "yamllint>=1.32.0,<2",
    { include-group = "types" },
]
mypy = [
    "django-stubs[compatible-mypy]==5.2.*",
    { include-group = "types" },
]
types = [
    "types-bleach>=6.2.0,<7",
    "types-paramiko>=3.5.0,<5",
    "types-Markdown~=3.7",
    "types-PyYAML>=6.0.12,<7",
    "celery-types",
]

[tool.uv]

[tool.uv.sources]
django-yamlfield = { git = "https://github.com/bakatrouble/django-yamlfield.git", rev = "c92d0373d12a02d1e52fb09b44010f156111d7ea" }

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.sdist]
include = [
    "course",
    "accounts",
    "relate",
]

[tool.hatch.build.targets.wheel]
include = [
    "course",
    "accounts",
    "relate",
]

[tool.ruff]
preview = true
target-version = "py310"
exclude = ["contrib/jupyterlite"]
include = ["*.py", "docker-image-run-py/runcode"]

[tool.ruff.lint]
extend-select = [
    "B",   # flake8-bugbear
    "C",   # flake8-comprehensions
    "E",   # pycodestyle
    "F",   # pyflakes
    "I",   # flake8-isort
    "N",   # pep8-naming
    "NPY", # numpy
    "Q",   # flake8-quotes
    "W",   # pycodestyle
    "RUF",
    "UP",
    "TC",
    # "DJ",
    "FURB",
]

extend-ignore = [
    "C90",  # McCabe complexity
    "E221", # multiple spaces before operator
    "E226", # missing whitespace around arithmetic operator
    "E241", # multiple spaces after comma
    "E242", # tab after comma
    "E402", # module level import not at the top of file
    "N818", # error suffix in exception names

    # TODO
    "B904", # raise in except needs from
    "B028", # stacklevel
    "RUF012", # mutable class atttributes
    "UP031", # %-format
]
allowed-confusables = ["‐", "–"]

[tool.ruff.lint.flake8-quotes]
docstring-quotes = "double"
inline-quotes = "double"
multiline-quotes = "double"

[tool.ruff.lint.per-file-ignores]
"course/mdx_mathjax.py" = ["N802"]

# config file, no type annotations, avoid scaring users
"local_settings.py" = ["I002"]

# config file, no type annotations, avoid scaring users
"local_settings_example.py" = ["I002"]

# copy-pasted from elsewhere, not our problem
"saml-config/attribute-maps/*.py" = ["Q", "E231", "W292", "W291", "E501", "I002"]

# mostly generated
"*/migrations/*.py" = ["Q", "E501", "RUF012", "F401", "I002", "UP", "N813"]

# TODO
"tests/**/*.py" = ["F841", "RUF012"]

[tool.ruff.lint.isort]
combine-as-imports = true
lines-after-imports = 2
required-imports = ["from __future__ import annotations"]

[tool.mypy]
plugins = ["mypy_django_plugin.main"]

strict_optional = true
ignore_missing_imports = true
disallow_untyped_calls = "true"
# wishful thinking :)
# disallow_untyped_defs= "true"
show_error_codes = true
untyped_calls_exclude = [
    "dulwich",
]
allow_untyped_calls = true
allow_redefinition_new = true
local_partial_types = true

[tool.django-stubs]
django_settings_module = "relate.settings"
strict_settings = false

[[tool.mypy.overrides]]
module = [
    "course.migrations.*",
]
ignore_errors = true

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "relate.settings"
python_files = [
    "tests.py",
    "test_*.py",
    "*_tests.py",
]


markers = [
    "slow: marks tests as slow (run with '--slow')",
    "postgres: mark test as postgres specific",
]

[tool.typos.default]
extend-ignore-re = [
  "(?Rm)^.*(#|//)\\s*spellchecker:\\s*disable-line"
]

[tool.typos.default.extend-words]
# opps as in 'grading opportunities'
opps = "opps"
# fre, short for flow rule exception
fre = "fre"
# short for 'stipulation'
stip = "stip"
# as in documen.tician.de
documen = "documen"
# like sur_name in SAML
sur = "sur"
# like serialization
"ser" = "ser"

[tool.typos.files]
extend-exclude = [
"locale/**/*.po",
# migrations reflect the past, don't attempt to alter them.
"*/migrations/**/*.py",
# not ours
"saml-config/**/*.py",
"contrib/jupyterlite/mamba-root",
]

[tool.basedpyright]
reportImplicitStringConcatenation = "none"
reportUnnecessaryIsInstance = "none"
reportUnusedCallResult = "none"
reportExplicitAny = "none"
reportUnreachable = "hint"

# This reports even cycles that are qualified by 'if TYPE_CHECKING'. Not what
# we care about at this moment.
# https://github.com/microsoft/pyright/issues/746
reportImportCycles = "none"
pythonVersion = "3.11"
pythonPlatform = "All"

reportAny = "none"

exclude = [
    ".venv",
    "bulk-storage",
    "git-roots",
    "node_modules",
    "course/migrations",
    "accounts/migrations",
    "prairietest/migrations",
    "local_settings.py",
]

[[tool.basedpyright.executionEnvironments]]
root = "tests"
extraPaths = ["."]

reportAny = "none"
reportArgumentType = "hint"
reportAssignmentType = "none"
reportAttributeAccessIssue = "hint"
reportCallIssue = "hint"
reportConstantRedefinition = "none"
reportGeneralTypeIssues = "hint"
reportIncompatibleMethodOverride = "none"
reportIncompatibleVariableOverride = "none"
reportIndexIssue = "hint"
reportMissingTypeArgument = "hint"
reportOptionalIterable = "none"
reportOptionalMemberAccess = "none"
reportOptionalOperand = "none"
reportOptionalSubscript = "none"
reportPossiblyUnboundVariable = "none"
reportPrivateUsage = "none"
reportUninitializedInstanceVariable = "hint"
reportUnknownArgumentType = "hint"

# vim: foldmethod=marker
